Move the argsparser into extensions folder
